primevape-backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-CORS==4.0.0
Flask-JWT-Extended==4.6.0
Flask-Bcrypt==1.0.1
python-dotenv==1.0.0
Werkzeug==3.0.1
SQLAlchemy>=2.0.35
email-validator>=2.3.0
